void rain() {   rainTotal = float(((nyb(18) * 100000) + (nyb(17) * 10000) + (nyb(16) * 1000) + (nyb(15) * 100) + (nyb(14) * 10) + nyb(13)) * 25.4) / 1000.0;
Rob, after doing some extensive testing I came to the conclusion that your Rain calculations should multiply by 25.4 (instead of 25). This proved to be 100% accurate and yield exactly the same results as the Oregon base station. I am not sure, but I think it has to do with the fact that there are 25.4 mm in an inch?